Add more theme tweaks
